+++ CHANGELOG.md
@@ -12,6 +12,8 @@ CHANGELOG for FlatCAM beta
 - remade the Excellon Object Properties UI
 - remade the Geometry Object Properties UI
 - remade the CNC Job Object Properties UI
+- in Etch Compensation Plugin updated the GUI
+- small UI updates in other Plugins
 
 5.09.2021